chore(api/db): recreate tables only when running dev

+++ services/api/src/api/db/connection.py
@@ -1,6 +1,6 @@
 from typing import Annotated, TypeVar
 
-from api.common.utils.env import genv
+from api.db.config import create_db_url
 from fastapi.params import Depends
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from sqlmodel import SQLModel
@@ -10,20 +10,11 @@ T = TypeVar("T")
 _engine: AsyncEngine | None = None
 
 
-def create_db_url():
-    user = genv("POSTGRES_USER", "app")
-    password = genv("POSTGRES_PASSWORD", "1234")
-    host = genv("POSTGRES_HOST", "127.0.0.1")
-    port = genv("POSTGRES_PORT", "5432")
-    database = genv("POSTGRES_DB", "general")
-
-    return f"postgresql+asyncpg://{user}:{password}@{host}:{port}/{database}"
-
-
-async def init():
-    engine = create_async_engine(create_db_url(), echo=True)
+async def init(debug: bool):
+    engine = create_async_engine(create_db_url(), echo=debug)
     async with engine.begin() as conn:
-        await conn.run_sync(SQLModel.metadata.drop_all)
+        if debug:
+            await conn.run_sync(SQLModel.metadata.drop_all)
         await conn.run_sync(SQLModel.metadata.create_all)
     global _engine
     _engine = engine
